remember time allocations
This will be used to find a good size of the number of move overheads to
remember. When we are under time pressure, the move frequency is higher,
and we should aim at being able to find the median move overheads for
the last 30 seconds or so. But if move times are higher, for example a
minute, we should store at most two samples.